+++ Sources/VComponents/Components/Containers/Dynamic Pager Tab View/VDynamicPagerTabViewState.swift
@@ -1,0 +1,17 @@
+//
+//  VStretchedIndicatorStaticPagerTabViewState.swift
+//  VComponents
+//
+//  Created by Vakhtang Kontridze on 01.09.23.
+//
+
+import Foundation
+import VCore
+
+// MARK: - V Dynamic Pager Tab View Tab Item Internal State
+/// Enumeration that represents state, such as `enabled`, `pressed`, or `disabled`.
+@available(iOS 14.0, *)
+@available(macOS 11.0, *)@available(macOS, unavailable)
+@available(tvOS 14.0, *)@available(tvOS, unavailable)
+@available(watchOS 8.0, *)@available(watchOS, unavailable)
+public typealias VDynamicPagerTabViewTabItemInternalState = GenericState_EnabledPressedDisabled